runmirrors
change the runmirrors.mirror format. We keep compatibility with the old format, so if one
uses them they should continue to do so.

If one uses the new style, one is able to add all possible ssh options one wants. They just have
to start with -. All that in there will be passed as-is to ssh.

Signed-off-by: Joerg Jaspert <joerg@debian.org>